@@ -12,44 +12,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.4.5",
-    "@angular/compiler": "2.4.5",
-    "@angular/compiler-cli": "2.4.5",
-    "@angular/core": "2.4.5",
-    "@angular/forms": "2.4.5",
-    "@angular/http": "2.4.5",
-    "@angular/platform-browser": "2.4.5",
-    "@angular/platform-browser-dynamic": "2.4.5",
-    "@angular/platform-server": "2.4.5",
-    "@angular/router": "3.4.5",
-    "@angular/tsc-wrapped": "^0.5.0",
-    "@angular/upgrade": "2.4.5",
-    "@types/lodash": "4.14.92",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.0.1",
-    "ts-helpers": "^1.1.1",
-    "angular2-jwt": "0.1.28",
-    "zone.js": "^0.7.2",
-    "ajv": "^5.0.0",
-    "angular2-json-schema-form": "0.3.0-alpha.26"
+    "@angular/cdk": "^5.0.4",
+    "@angular/animations": "^5.2.0",
+    "@angular/common": "^5.2.0",
+    "@angular/compiler": "^5.2.0",
+    "@angular/core": "^5.2.0",
+    "@angular/forms": "^5.2.0",
+    "@angular/http": "^5.2.0",
+    "@angular/material": "^5.0.4",
+    "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
+    "@angular/platform-server": "^5.2.0",
+    "@angular/router": "^5.2.0",
+    "@angular/flex-layout": "2.0.0-beta.12",
+    "ajv": "^6.0.1",
+    "@types/lodash": "^4.14.92",
+    "hammerjs": "2.x",
+    "angular2-json-schema-form": "0.7.0-alpha.1",
+    "angular2-jwt": "0.2.3",
+    "core-js": "^2.5.3",
+    "rxjs": "^5.5.6",
+    "ts-helpers": "^1.1.2",
+    "json2typescript": "~1.0.5"
   },
   "devDependencies": {
-    "@angular-devkit/schematics": "0.0.46",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "^6.0.42",
-    "@angular/cli": "1.6.4",
-    "angular-ide": "^0.9.17",
-    "codelyzer": "~2.0.0-beta.1",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "^1.3.0",
-    "karma-chrome-launcher": "^2.0.0",
+    "@angular/cli": "^1.6.4",
+    "@angular/compiler-cli": "^5.2.0",
+    "@types/jasmine": "2.8.3",
+    "@types/node": "^9.3.0",
+    "angular-ide": "^0.9.39",
+    "codelyzer": "~4.0.2",
+    "hammerjs": "2.x",
+    "jasmine-core": "2.8.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "2.0.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "~4.0.13",
-    "ts-node": "1.2.1",
-    "tslint": "^4.3.0",
-    "typescript": "~2.4.2"
+    "karma-jasmine": "^1.1.1",
+    "karma-remap-istanbul": "^0.6.0",
+    "protractor": "~5.2.2",
+    "ts-node": "4.1.0",
+    "tslint": "^5.9.1",
+    "typescript": "~2.4.2",
+    "zone.js": "^0.8.20"
   }
 }
