@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(value = "/platform/api/management/services")
 public class NmServiceDeploymentAdminRestController {
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     @RequestMapping(value = "", method = RequestMethod.GET)
     public String basicInfo() {
         return "This is NMaaS Platform REST API for NM Services configuration";
