@@ -45,7 +45,7 @@ public class AppDeploymentMonitorRestController {
      * @return list of deployments
      */
     @RequestMapping(value = "", method = RequestMethod.GET)
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     public List<AppDeploymentView> listAllDeployments() {
         return deploymentMonitor.allDeployments().stream()
                 .map(d -> modelMapper.map(d, AppDeploymentView.class))
@@ -59,7 +59,7 @@ public class AppDeploymentMonitorRestController {
      * @return current deployment state
      * @throws InvalidDeploymentIdException if deployment with provided identifier doesn't exist in the system
      */
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     @RequestMapping(value = "{deploymentId}/state", method = RequestMethod.GET)
     public AppLifecycleState loadDeploymentState(
             @PathVariable String deploymentId) throws InvalidDeploymentIdException {
@@ -74,7 +74,7 @@ public class AppDeploymentMonitorRestController {
      * @throws InvalidDeploymentIdException if deployment with provided identifier doesn't exist in the system
      * @throws InvalidAppStateException if deployment didn't complete yet
      */
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     @RequestMapping(value = "{deploymentId}/access", method = RequestMethod.GET)
     public AppUiAccessDetails loadDeploymentUserAccessInfo(
             @PathVariable String deploymentId) throws InvalidDeploymentIdException, InvalidAppStateException {
