@@ -32,7 +32,7 @@ public class AppLifecycleManagerRestController {
      * @param applicationId identifier of the application
      * @return unique identifier of the application deployment
      */
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     @RequestMapping(value = "",
             params = {"clientid", "applicationid"},
             method = RequestMethod.POST)
@@ -50,7 +50,7 @@ public class AppLifecycleManagerRestController {
      * @param configuration initial application configuration provided by the user/client
      * @throws InvalidDeploymentIdException if deployment with provided identifier doesn't exist in the system
      */
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     @RequestMapping(value = "/{deploymentId}",
             method = RequestMethod.POST,
             consumes = "application/json")
@@ -67,7 +67,7 @@ public class AppLifecycleManagerRestController {
      * @param deploymentId unique identifier of the application deployment
      * @throws InvalidDeploymentIdException if deployment with provided identifier doesn't exist in the system
      */
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_SUPERADMIN')")
     @RequestMapping(value = "/{deploymentId}",
             method = RequestMethod.DELETE)
     @ResponseStatus(code = HttpStatus.OK)
