@@ -16,11 +16,14 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.filter.CharacterEncodingFilter;
 
+import net.geant.nmaas.portal.persistent.entity.Domain;
 import net.geant.nmaas.portal.persistent.entity.Role;
 import net.geant.nmaas.portal.persistent.entity.User;
+import net.geant.nmaas.portal.persistent.repositories.DomainRepository;
 import net.geant.nmaas.portal.persistent.repositories.UserRepository;
+import net.geant.nmaas.portal.service.DomainService;
 import net.geant.nmaas.portal.service.FileStorageService;
-import net.geant.nmaas.portal.service.LocalFileStorageService;
+import net.geant.nmaas.portal.service.impl.LocalFileStorageService;
 
 @Configuration
 @ComponentScan(basePackages={"net.geant.nmaas.portal.service"})
@@ -36,24 +39,24 @@ public class PortalConfig {
 			@Autowired
 			private UserRepository userRepository;
 			
+			@Autowired
+			private DomainService domains;
+			
 			@Override
 			@Transactional
 			public void afterPropertiesSet() {
+				domains.createGlobalDomain();				
+				
 				Optional<User> admin = userRepository.findByUsername("admin");
 				if(!admin.isPresent())
-					addUser("admin", "admin", new Role[] { Role.ADMIN, Role.MANAGER, Role.USER });
+					addUser("admin", "admin", Role.ROLE_SUPERADMIN);
 			}
 
 			private void addUser(String username, String password, Role role) {								
-				User user = new User(username, passwordEncoder.encode(password), role);
+				User user = new User(username, passwordEncoder.encode(password), domains.getGlobalDomain(), role);
 				userRepository.save(user);
 			}
-			
-			private void addUser(String username, String password, Role[] roles) {								
-				User user = new User(username, passwordEncoder.encode(password), new ArrayList<Role>(Arrays.asList(roles)));
-				userRepository.save(user);
-			}
-			
+						
 		};
 	}
 	
