@@ -6,16 +6,19 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class UserSignup {
 	private String username;
 	
-	private String password;
-
+	private String password;	
+	
+	private Long domainId;
+	
 	protected UserSignup() {
 		
 	}
 	
 	@JsonCreator
-	public UserSignup(@JsonProperty(value="username", required=true) String username, @JsonProperty(value = "password", required=true) String password) {
+	public UserSignup(@JsonProperty(value="username", required=true) String username, @JsonProperty(value = "password", required=false) String password, @JsonProperty(value="domainId", required = false) Long domainId) {
 		this.username = username;
 		this.password = password;
+		this.domainId = domainId;
 	}
 
 	public String getUsername() {
@@ -25,6 +28,11 @@ public class UserSignup {
 	public String getPassword() {
 		return password;
 	}
+
+	public Long getDomainId() {
+		return domainId;
+	}
+	
 	
 	
 }
