@@ -16,6 +16,8 @@ import net.geant.nmaas.portal.persistent.entity.Application;
 import net.geant.nmaas.portal.persistent.entity.User;
 import net.geant.nmaas.portal.persistent.repositories.AppInstanceRepository;
 import net.geant.nmaas.portal.persistent.repositories.UserRepository;
+import net.geant.nmaas.portal.service.DomainService;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -31,7 +33,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping("/portal/api/apps/instances")
+@RequestMapping("/portal/api")
 public class AppInstanceController extends AppBaseController {
 
 	@Autowired
@@ -48,27 +50,56 @@ public class AppInstanceController extends AppBaseController {
 
 	@Autowired
 	ModelMapper modelMapper;
+	
+	@Autowired
+	DomainService domains;
 
-	@RequestMapping(method = RequestMethod.GET)
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	public List<AppInstance> getAllInstances(Pageable pageable) {
+	@GetMapping("/apps/instances")
+	@PreAuthorize("hasRole('ROLE_SUPERADMIN')")
+	public List<AppInstance> getAllInstances(Pageable pageable) throws MissingElementException {				
 		return appInstanceRepo.findAll(pageable).getContent().stream().map(appInstance -> mapAppInstance(appInstance)).collect(Collectors.toList());
 	}
 
-	@RequestMapping(value = "my", method = RequestMethod.GET)
-	public List<AppInstance> getMyAllInstances(@NotNull Principal principal, Pageable pageable)
-			throws MissingElementException {
-		return getUserAppInstances(principal.getName(), pageable);
+	@GetMapping("/apps/instances/my")
+	public List<AppInstance> getMyAllInstances(@NotNull Principal principal, Pageable pageable) throws MissingElementException {
+		Optional<User> user = userRepo.findByUsername(principal.getName());
+		if (!user.isPresent())
+			throw new MissingElementException("User not found");
+
+		return appInstanceRepo.findAllByOwner(user.get(), pageable).getContent().stream().map(appInstance -> mapAppInstance(appInstance)).collect(Collectors.toList());
+	}
+	
+	
+	@GetMapping("/domains/{domainId}/apps/instances")
+	@PreAuthorize("hasPermission(#domainId, 'domain', 'OWNER')")
+	public List<AppInstance> getAllInstances(@PathVariable Long domainId, Pageable pageable) throws MissingElementException {
+		net.geant.nmaas.portal.persistent.entity.Domain domain = domains.findDomain(domainId);
+		if(domain == null)
+			throw new MissingElementException("Domain not found");
+
+		return appInstanceRepo.findAllByDomain(domain, pageable).getContent().stream().map(appInstance -> mapAppInstance(appInstance)).collect(Collectors.toList());
 	}
 
-	@RequestMapping(value = "/user/{username}", method = RequestMethod.GET)
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	public List<AppInstance> getUserAllInstances(@PathVariable("username") String username, Pageable pageable)
+	@GetMapping(value = "/domains/{domainId}/apps/instances/my")
+	@PreAuthorize("hasPermission(#domainId, 'domain', 'ANY')")
+	public List<AppInstance> getMyAllInstances(@PathVariable Long domainId, @NotNull Principal principal, Pageable pageable)
 			throws MissingElementException {
-		return getUserAppInstances(username, pageable);
+		return getUserDomainAppInstances(domainId, principal.getName(), pageable);
 	}
 
-	private List<AppInstance> getUserAppInstances(String username, Pageable pageable) throws MissingElementException {
+	@GetMapping("/domains/{domainId}/apps/instances/user/{username}")
+	@PreAuthorize("hasPermission(#domainId, 'domain', 'OWNER')")
+	public List<AppInstance> getUserAllInstances(@PathVariable Long domainId, @PathVariable String username, Pageable pageable)
+			throws MissingElementException {
+		return getUserDomainAppInstances(domainId, username, pageable);
+	}
+
+	private List<AppInstance> getUserDomainAppInstances(Long domainId, String username, Pageable pageable) throws MissingElementException {
+		
+		net.geant.nmaas.portal.persistent.entity.Domain domain = domains.findDomain(domainId);
+		if(domain == null)
+			throw new MissingElementException("Domain not found");
+		
 		Optional<User> user = userRepo.findByUsername(username);
 		if (!user.isPresent())
 			throw new MissingElementException("User not found");
@@ -76,8 +107,8 @@ public class AppInstanceController extends AppBaseController {
 		return appInstanceRepo.findAllByOwner(user.get(), pageable).getContent().stream().map(appInstance -> mapAppInstance(appInstance)).collect(Collectors.toList());
 	}
 	
-	@RequestMapping(value = "/{appInstanceId}", method = RequestMethod.GET)
-	@PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#appInstanceId, 'appInstance', 'owner')")
+	@GetMapping({"/apps/instances/{appInstanceId}", "/domains/{domainId}/apps/instances/{appInstanceId}"})
+	@PreAuthorize("hasPermission(#appInstanceId, 'appInstance', 'OWNER')")
 	public AppInstance getAppInstance(@PathVariable(value = "appInstanceId") Long appInstanceId,
 			@NotNull Principal principal) throws MissingElementException, ProcessingException {
 		net.geant.nmaas.portal.persistent.entity.AppInstance appInstance = appInstanceRepo.findOne(appInstanceId);
@@ -87,15 +118,21 @@ public class AppInstanceController extends AppBaseController {
 		return mapAppInstance(appInstance);
 	}
 
-	@RequestMapping(method = RequestMethod.POST)
+	@PostMapping("/domains/{domainId}/apps/instances")
+	@PreAuthorize("hasPermission(#domainId, 'domain', 'CREATE')")
 	@Transactional
 	public Id createAppInstance(@RequestBody(required = true) AppInstanceSubscription appInstanceSubscription,
-			@NotNull Principal principal) throws MissingElementException {
+			@NotNull Principal principal, @PathVariable Long domainId) throws MissingElementException {
 		Application app = getApp(appInstanceSubscription.getApplicationId());
 		User user = getUser(principal.getName());
 
 		net.geant.nmaas.portal.persistent.entity.AppInstance appInstance = new net.geant.nmaas.portal.persistent.entity.AppInstance(
 				app, appInstanceSubscription.getName());
+		
+		net.geant.nmaas.portal.persistent.entity.Domain domain = domains.findDomain(domainId);
+		if(domain == null)
+			throw new MissingElementException("Domain not found");
+		appInstance.setDomain(domain);
 
 		Identifier internalId = appLifecycleManager.deployApplication(new Identifier(Long.toString(user.getId())),
 				new Identifier(Long.toString(app.getId())));
@@ -106,8 +143,9 @@ public class AppInstanceController extends AppBaseController {
 		return new Id(appInstance.getId());
 	}
 
-	@RequestMapping(value = "/{appInstanceId}", method = RequestMethod.DELETE)
-	@PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#appInstanceId, 'appInstance', 'owner')")
+	
+	@DeleteMapping({"/apps/instances/{appInstanceId}", "/domains/{domainId}/apps/instances/{appInstanceId}"})
+	@PreAuthorize("hasPermission(#appInstanceId, 'appInstance', 'DELETE')")
 	@Transactional
 	public void deleteAppInstance(@PathVariable(value = "appInstanceId") Long appInstanceId,
 			@NotNull Principal principal) throws MissingElementException, ProcessingException {
@@ -120,8 +158,8 @@ public class AppInstanceController extends AppBaseController {
 		}
 	}
 
-	@RequestMapping(value = "/{appInstanceId}/configure", method = RequestMethod.POST)
-	@PreAuthorize("hasPermission(#appInstanceId, 'appInstance', 'owner')")
+	@PostMapping({"/apps/instances/{appInstanceId}/configure", "/domains/{domainId}/apps/instances/{appInstanceId}/configure"})
+	@PreAuthorize("hasPermission(#appInstanceId, 'appInstance', 'OWNER')")
 	@Transactional
 	public void applyConfiguration(@PathVariable(value = "appInstanceId") Long appInstanceId,
 			@RequestBody String configuration, @NotNull Principal principal)
@@ -146,8 +184,8 @@ public class AppInstanceController extends AppBaseController {
 		}
 	}
 
-	@RequestMapping(value = "/{appInstanceId}/state", method = RequestMethod.GET)
-	@PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#appInstanceId, 'appInstance', 'owner')")
+	@GetMapping("/domains/{domainId}/apps/instances/{appInstanceId}/state")
+	@PreAuthorize("hasPermission(#appInstanceId, 'appInstance', 'OWNER')")
 	public AppInstanceStatus getState(@PathVariable(value = "appInstanceId") Long appInstanceId,
 			@NotNull Principal principal) throws MissingElementException, ProcessingException {
 		net.geant.nmaas.portal.persistent.entity.AppInstance appInstance = getAppInstance(appInstanceId);
@@ -261,9 +299,13 @@ public class AppInstanceController extends AppBaseController {
 	}
 
 	private AppInstance mapAppInstance(net.geant.nmaas.portal.persistent.entity.AppInstance appInstance) {
+		if(appInstance == null)
+			return null;
+		
 		AppInstance ai = modelMapper.map(appInstance, AppInstance.class);
 		try {
 			ai.setState(mapAppInstanceState(this.appDeploymentMonitor.state(appInstance.getInternalId())));
+			ai.setDomainId(appInstance.getDomain().getId());
 		} catch (Exception e) {
 			ai.setState(AppInstanceState.UNKNOWN);
 			ai.setUrl(null);
