@@ -1,26 +1,27 @@
 package net.geant.nmaas.portal.api.security;
 
 import java.io.Serializable;
+import java.util.HashSet;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Configurable;
 import org.springframework.security.access.PermissionEvaluator;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
-import net.geant.nmaas.portal.persistent.entity.AppInstance;
-import net.geant.nmaas.portal.persistent.entity.Comment;
-import net.geant.nmaas.portal.persistent.repositories.AppInstanceRepository;
-import net.geant.nmaas.portal.persistent.repositories.CommentRepository;
+import net.geant.nmaas.portal.persistent.entity.User;
+import net.geant.nmaas.portal.service.AclService;
+import net.geant.nmaas.portal.service.AclService.Permissions;
+import net.geant.nmaas.portal.service.UserService;
 
 @Component
 public class ApiPermissionEvaluator implements PermissionEvaluator {
-
-	@Autowired
-	private CommentRepository commentRepository;
 	
 	@Autowired
-	private AppInstanceRepository appInstanceRepository;
+	private UserService users;
+	
+	@Autowired
+	AclService aclService;
 	
 	public ApiPermissionEvaluator() {
 		
@@ -34,53 +35,40 @@ public class ApiPermissionEvaluator implements PermissionEvaluator {
 	@Override
 	public boolean hasPermission(Authentication authentication, Serializable targetId, String targetType,
 			Object permission) {
-		
-		if(authentication == null || authentication.getName() == null || targetId == null || targetType == null || permission == null || !(permission instanceof String))
+				
+		if(authentication == null || authentication.getName() == null || targetType == null || permission == null || !(permission instanceof String))
 			return false;
-		
+						
 		String targetName;
-		String permissionName = (String)permission;
+		String permissionStr = (String)permission;
 		
 		if(targetType instanceof String)
 			targetName = targetType;
 		else
 			targetName = targetType.getClass().getSimpleName();
 		targetName = targetName.toLowerCase();
+								
+		User user = users.findByUsername(authentication.getName());
+		if(user == null)
+			return false;
 		
-		permissionName = permissionName.toLowerCase();
+		Permissions[] perms = convertToPermissions(permissionStr);
 		
-		switch(targetName) {
-			case "comment": if(!(targetId instanceof Long))
-								return false;
-							else
-								return hasPermissionComment(authentication, (Long) targetId, permissionName);
-			case "appinstance": if(!(targetId instanceof Long))
-									return false;
-								else
-									return hasPermissionAppInstance(authentication, (Long) targetId, permissionName);
-		}
-		
-		return false;
+		return aclService.isAuthorized(user.getId(), targetId, targetType, perms);		
 	}
 
-	private boolean hasPermissionComment(Authentication authentication, Long commentId, String permission) {
-		Comment comment = commentRepository.findOne(commentId);
-		switch(permission) {
-			case "owner":
-				return (comment != null && comment.getOwner() != null && authentication.getName().equals(comment.getOwner().getUsername()));
+	private Permissions[] convertToPermissions(String permissionStr) {
+		
+		Set<Permissions> perms = new HashSet<Permissions>();
+		
+		String[] permArray = (permissionStr != null ? permissionStr.trim().split(",") : null); 
+		if(permArray != null) {
+			for(String perm : permArray)
+				if(perm != null || perm.trim().length() > 0)
+					perms.add(Permissions.valueOf(perm));
 		}
 		
-		return false;
-	}
-
-	private boolean hasPermissionAppInstance(Authentication authentication, Long appInstanceId, String permission) {
-		AppInstance appInstance = appInstanceRepository.findOne(appInstanceId);
-		switch(permission) {
-			case "owner":
-				return (appInstance != null && appInstance.getOwner() != null && authentication.getName().equals(appInstance.getOwner().getUsername()));
-		}
-		
-		return false;
+		return perms.toArray(new Permissions[perms.size()]);
 	}
 
 }
