@@ -32,7 +32,7 @@ public class DockerHostManagerApiSecurityTest extends BaseControllerTest {
 
     @Test
     public void shouldAuthorizeAdminProperUser() throws Exception {
-        String token = getValidUserTokenFor(Role.ADMIN);
+        String token = getValidUserTokenFor(Role.ROLE_SUPERADMIN);
         mvc.perform(get("/platform/api/management/dockerhosts")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isOk());
@@ -40,7 +40,7 @@ public class DockerHostManagerApiSecurityTest extends BaseControllerTest {
 
     @Test
     public void shouldRejectNonAdminProperUser() throws Exception {
-        String token = getValidUserTokenFor(Role.USER);
+        String token = getValidUserTokenFor(Role.ROLE_USER);
         mvc.perform(get("/platform/api/management/dockerhosts")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isUnauthorized());
