@@ -57,7 +57,7 @@ public class NmServiceConfigurationTemplateAdminRestControllerTest extends BaseC
 
     @Test
     public void shouldStoreAndLoadTemplate() throws Exception {
-        String token = getValidUserTokenFor(Role.ADMIN);
+        String token = getValidUserTokenFor(Role.ROLE_SUPERADMIN);
         mvc.perform(post("/platform/api/management/configurations/templates")
                 .header("Authorization", "Bearer " + token)
                 .contentType(MediaType.APPLICATION_JSON)
@@ -72,7 +72,7 @@ public class NmServiceConfigurationTemplateAdminRestControllerTest extends BaseC
 
     @Test
     public void shouldAuthAndForbidSimpleGet() throws Exception {
-        String token = getValidUserTokenFor(Role.USER);
+        String token = getValidUserTokenFor(Role.ROLE_USER);
         mvc.perform(get("/platform/api/management/configurations/templates")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isUnauthorized());
