@@ -41,7 +41,7 @@ public class NmServiceDeploymentAdminApiSecurityTest extends BaseControllerTest 
 
     @Test
     public void shouldAuthAndCallSimpleGet() throws Exception {
-        String token = getValidUserTokenFor(Role.ADMIN);
+        String token = getValidUserTokenFor(Role.ROLE_SUPERADMIN);
         mvc.perform(get("/platform/api/management/services")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isOk());
@@ -49,7 +49,7 @@ public class NmServiceDeploymentAdminApiSecurityTest extends BaseControllerTest 
 
     @Test
     public void shouldAuthAndForbidSimpleGet() throws Exception {
-        String token = getValidUserTokenFor(Role.USER);
+        String token = getValidUserTokenFor(Role.ROLE_USER);
         mvc.perform(get("/platform/api/management/services")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isUnauthorized());
