@@ -69,7 +69,7 @@ public class KubernetesTemplateAdminRestControllerTest extends BaseControllerTes
     @Test
     @Transactional
     public void shouldStoreAndLoadTemplate() throws Exception {
-        String token = getValidUserTokenFor(Role.ADMIN);
+        String token = getValidUserTokenFor(Role.ROLE_SUPERADMIN);
         assertThat(templateRepository.count(), equalTo(0L));
         mvc.perform(post("/platform/api/management/apps/{appId}/kubernetes/template", appId)
                 .header("Authorization", "Bearer " + token)
@@ -91,7 +91,7 @@ public class KubernetesTemplateAdminRestControllerTest extends BaseControllerTes
     @Test
     @Transactional
     public void shouldReturnProperCodesOnExceptions() throws Exception {
-        String token = getValidUserTokenFor(Role.ADMIN);
+        String token = getValidUserTokenFor(Role.ROLE_SUPERADMIN);
         mvc.perform(get("/platform/api/management/apps/{appId}/kubernetes/template", 100)
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isNotFound());
@@ -109,7 +109,7 @@ public class KubernetesTemplateAdminRestControllerTest extends BaseControllerTes
 
     @Test
     public void shouldAuthAndForbidSimpleGet() throws Exception {
-        String token = getValidUserTokenFor(Role.USER);
+        String token = getValidUserTokenFor(Role.ROLE_USER);
         mvc.perform(get("/platform/api/management/apps/{appId}/kubernetes/template", appId)
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isUnauthorized());
