@@ -33,7 +33,7 @@ public class OrchestratorApiSecurityTest extends BaseControllerTest {
 
     @Test
     public void shouldAuthorizeAdminProperUser() throws Exception {
-        String token = getValidUserTokenFor(Role.ADMIN);
+        String token = getValidUserTokenFor(Role.ROLE_SUPERADMIN);
         mvc.perform(get("/platform/api/orchestration/deployments")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isOk());
@@ -45,7 +45,7 @@ public class OrchestratorApiSecurityTest extends BaseControllerTest {
 
     @Test
     public void shouldRejectNonAdminProperUser() throws Exception {
-        String token = getValidUserTokenFor(Role.USER);
+        String token = getValidUserTokenFor(Role.ROLE_USER);
         mvc.perform(get("/platform/api/orchestration/deployments")
                 .header("Authorization", "Bearer " + token))
                 .andExpect(status().isUnauthorized());
