@@ -1,50 +1,60 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from "@angular/router";
-import { CommonModule } from '@angular/common';
-import { FormsModule }   from '@angular/forms';
+import {NgModule} from '@angular/core';
+import {RouterModule} from '@angular/router';
+import {CommonModule} from '@angular/common';
+import {FormsModule} from '@angular/forms';
 
-import { AuthModule } from '../auth/auth.module';
+import {AuthModule} from '../auth/auth.module';
 
-import { AppMarketComponent } from './appmarket.component';
-import { AppListModule } from './applist/applist.module';
-import { AppDetailsComponent } from './appdetails/index';
-import { AppInstanceModule } from './appinstance/appinstance.module';
+import {AppMarketComponent} from './appmarket.component';
+import {AppListModule} from './applist/applist.module';
+import {AppDetailsComponent} from './appdetails/index';
+import {AppInstanceModule} from './appinstance/appinstance.module';
 
-import { NavbarComponent } from './navbar/index';
-import { LogoutComponent } from '../logout/index';
+import {NavbarComponent} from './navbar/index';
+import {LogoutComponent} from '../logout/index';
 
-import { SharedModule } from '../shared/shared.module';
+import {SharedModule} from '../shared/shared.module';
 
-import { AppsService } from '../service/apps.service';
-import { TagService } from '../service/tag.service';
-import { AppInstallModalComponent } from './appinstall/appinstallmodal.component';
+import {AppsService} from '../service/apps.service';
+import {DomainService} from '../service/domain.service';
+import {TagService} from '../service/tag.service';
+import {UserService} from '../service/user.service';
 
-import { PipesModule } from '../pipe/pipes.module';
+import {AppInstallModalComponent} from './appinstall/appinstallmodal.component';
+
+import {PipesModule} from '../pipe/pipes.module';
+import {DomainsModule} from './domains/domains.module';
+import {UsersModule} from './users/users.module';
 
 @NgModule({
-    declarations: [
-        AppMarketComponent,
-        AppDetailsComponent,
-        NavbarComponent,
-        AppInstallModalComponent
-    ],
-    imports: [ 
-        FormsModule,
-        CommonModule,
-        RouterModule,
-        SharedModule,
-        AppListModule,
-        AppInstanceModule,
-        AuthModule,
-        PipesModule
-    ],
-    exports: [
-        AppMarketComponent
-    ],
-    providers: [
-        AppsService,
-        TagService
-    ]
-    
+  declarations: [
+    AppMarketComponent,
+    AppDetailsComponent,
+    NavbarComponent,
+    AppInstallModalComponent
+  ],
+  imports: [
+    FormsModule,
+    CommonModule,
+    RouterModule,
+    SharedModule,
+    AppListModule,
+    AppInstanceModule,
+    DomainsModule,
+    UsersModule,
+    AuthModule,
+    PipesModule
+  ],
+  exports: [
+    AppMarketComponent
+  ],
+  providers: [
+    AppsService,
+    DomainService,
+    UserService,
+    TagService,
+    UserService
+  ]
+
 })
-export class AppMarketModule { }
+export class AppMarketModule {}
