@@ -1,3 +1,3 @@
-export * from './appmarket.module'
-export * from './appmarket.component'
-export * from './appmarket.routes'
+export * from './appmarket.module';
+export * from './appmarket.component';
+export * from './appmarket.routes';
