@@ -1,24 +1,24 @@
-import { Injectable } from '@angular/core';
-import { Router } from '@angular/router';
-import { CanActivate } from '@angular/router';
-import { AuthService } from './auth.service';
+import {Injectable} from '@angular/core';
+import {Router} from '@angular/router';
+import {CanActivate} from '@angular/router';
+import {AuthService} from './auth.service';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
 
-    constructor(private auth: AuthService, private router: Router ) { }
+  constructor(private auth: AuthService, private router: Router) {}
 
-    public canActivate(): boolean {
-        if (this.auth.isLogged()) {
-            return true;
-        }
+  public canActivate(): boolean {
+    if (this.auth.isLogged()) {
+      return true;
+    }
 
-        // not logged in so redirect to login page
-        this.router.navigate(['/login']);
-        return false;
-    }
-    
-    public canActivateChild(): boolean {
-       return this.canActivate(); 
-    }
+    // not logged in so redirect to login page
+    this.router.navigate(['/login']);
+    return false;
+  }
+
+  public canActivateChild(): boolean {
+    return this.canActivate();
+  }
 }
