@@ -1,5 +1,13 @@
-export class Id { 
-   constructor(public id: Number) {
-   		this.id = id;
-   }
+import {JsonObject, JsonProperty} from 'json2typescript';
+
+@JsonObject
+export class Id {
+
+  @JsonProperty('id', Number)
+  public id: Number = undefined;
+  
+  constructor(id?: Number) {
+    this.id = id;
+  }
+
 }
