@@ -1,14 +1,14 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from "@angular/common";
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
 
-import { AuthModule } from "../auth/auth.module";
+import {AuthModule} from '../auth/auth.module';
 
-import { GroupPipe, SecurePipe, AuthHttpWrapper } from "./index";
+import {GroupPipe, SecurePipe, KeysPipe, AuthHttpWrapper} from './index';
 
 @NgModule({
-  declarations:[ GroupPipe, SecurePipe ],
-  imports:[ CommonModule, AuthModule ],
-  exports:[ GroupPipe, SecurePipe ],
-  providers: [ AuthHttpWrapper ]
+  declarations: [GroupPipe, SecurePipe, KeysPipe],
+  imports: [CommonModule, AuthModule],
+  exports: [GroupPipe, SecurePipe, KeysPipe],
+  providers: [AuthHttpWrapper]
 })
-export class PipesModule{}
+export class PipesModule {}
