@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http } from "@angular/http";
+import { Http } from '@angular/http';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 
@@ -20,6 +20,14 @@ export class AppConfigService {
     }
 
     public getApiUrl(): string {
-        return this.config.apiUrl;
+        return this.config.apiUrl || 'http://localhost/portal/api';
+    }
+
+    public getNmaasGlobalDomainId(): number {
+      return this.config.nmaas.globalDomainId || 0;
+    }
+
+    public getHttpTimeout(): number {
+      return this.config.http.timeout || 10000;
     }
 }
