@@ -25,7 +25,7 @@ export class AppsService {
         return this.authHttp.get(this.appConfig.getApiUrl() + '/apps')
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public getApp(id: Number): Observable<Application> {
@@ -33,48 +33,48 @@ export class AppsService {
         return this.authHttp.get(this.appConfig.getApiUrl() + '/apps/' + id)
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public getAppRateByUrl(urlPath: string): Observable<Rate> {
         return this.authHttp.get(this.appConfig.getApiUrl() + urlPath)
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public setMyAppRateByUrl(urlPath: string): Observable<any> {
         return this.authHttp.post(this.appConfig.getApiUrl() + urlPath, null)
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public getAppCommentsByUrl(urlPath: string): Observable<Comment[]> {
         return this.authHttp.get(this.appConfig.getApiUrl() + urlPath)
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public addAppCommentByUrl(urlPath: string, comment: Comment): Observable<Id> {
         return this.authHttp.post(this.appConfig.getApiUrl() + urlPath, comment)
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public deleteAppCommentByUrl(urlPath: string, id: Id): Observable<any> {
         return this.authHttp.delete(this.appConfig.getApiUrl() + urlPath + '/' + id.id)
             .timeout(10000)            
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
     
     public getAppScreenshotsByUrl(urlPath: string): Observable<FileInfo[]> {
         return this.authHttp.get(this.appConfig.getApiUrl() + urlPath)
             .timeout(10000)
             .map((res: Response) => res.json())
-            .catch((error: any) => Observable.throw((typeof error.json === "function" ? error.json().message : 'Server error')));
+            .catch((error: any) => Observable.throw((typeof error.json === 'function' ? error.json().message : 'Server error')));
     }
 
     public getAppScreenshotUrl(urlPath: string): string {
